refactor(tests): replace chunked rust test modules
Reorganize src/tests and src/server/tests into semantic modules.

Preserve shard manifest coverage and rust source line budget gates.
