+++ src/server/tests/admin_logs_and_summary.rs
@@ -1,3 +1,8 @@
+use super::*;
+use super::core_support_and_parsing::*;
+use super::linuxdo_oauth_and_admin_keys::*;
+use super::upstream_support_and_manual_jobs::*;
+
     #[test]
     fn session_delete_unsupported_views_render_neutral_result_status() {
         let token_record = TokenLogRecord {
