+++ src/server/tests/alerts_and_ha.rs
@@ -1,3 +1,8 @@
+use super::*;
+use super::core_support_and_parsing::*;
+use super::linuxdo_oauth_and_admin_keys::*;
+use super::upstream_support_and_manual_jobs::*;
+
 #[tokio::test]
 async fn alerts_endpoints_and_dashboard_recent_alerts_share_default_window() {
     let db_path = temp_db_path("alerts-dashboard-default-window");
