+++ src/server/tests/api_keys_and_registration.rs
@@ -1,3 +1,8 @@
+use super::*;
+use super::core_support_and_parsing::*;
+use super::linuxdo_oauth_and_admin_keys::*;
+use super::upstream_support_and_manual_jobs::*;
+
     #[tokio::test]
     async fn api_keys_batch_does_not_override_existing_group() {
         let db_path = temp_db_path("keys-batch-group-no-override");
