+++ src/server/tests/log_catalog_and_dashboard_sse.rs
@@ -1,3 +1,8 @@
+use super::*;
+use super::core_support_and_parsing::*;
+use super::linuxdo_oauth_and_admin_keys::*;
+use super::upstream_support_and_manual_jobs::*;
+
     #[tokio::test]
     async fn key_and_token_logs_catalog_scope_and_cache_ttl_work() {
         let db_path = temp_db_path("key-token-logs-catalog");
