+++ src/server/tests/mcp_billing_and_sessions.rs
@@ -1,3 +1,7 @@
+use super::*;
+use super::core_support_and_parsing::*;
+use super::upstream_support_and_manual_jobs::*;
+
     #[tokio::test]
     async fn mcp_batch_without_ids_search_and_research_charge_full_reserved_fallback_when_usage_missing()
      {
