+++ src/server/tests/system_settings_and_forward_proxy.rs
@@ -1,3 +1,7 @@
+use super::*;
+use super::core_support_and_parsing::*;
+use super::upstream_support_and_manual_jobs::*;
+
     #[tokio::test]
     async fn admin_system_settings_put_preserves_request_rate_limit_when_legacy_payload_omits_it() {
         let db_path = temp_db_path("admin-system-settings-legacy-request-rate");
