+++ src/server/tests/tavily_http_search.rs
@@ -1,3 +1,7 @@
+use super::*;
+use super::core_support_and_parsing::*;
+use super::upstream_support_and_manual_jobs::*;
+
     #[tokio::test]
     async fn admin_user_management_requires_admin() {
         let db_path = temp_db_path("admin-users-authz");
