+++ src/tests/proxy_affinity_and_summary.rs
@@ -1,3 +1,5 @@
+use super::*;
+
 #[tokio::test]
 async fn add_or_undelete_key_with_hint_only_proxy_affinity_refresh_invalidates_cached_record() {
     let db_path = temp_db_path("proxy-affinity-hint-cache-refresh");
@@ -1713,368 +1715,6 @@ async fn summary_quota_totals_exclude_quarantined_keys() {
     let _ = std::fs::remove_file(db_path);
 }
 
-async fn insert_summary_window_bucket(
-    proxy: &TavilyProxy,
-    key_id: &str,
-    bucket_start: i64,
-    total_requests: i64,
-    success_count: i64,
-    error_count: i64,
-    quota_exhausted_count: i64,
-) {
-    sqlx::query(
-        r#"
-        INSERT INTO api_key_usage_buckets (
-            api_key_id,
-            bucket_start,
-            bucket_secs,
-            total_requests,
-            success_count,
-            error_count,
-            quota_exhausted_count,
-            valuable_success_count,
-            valuable_failure_count,
-            other_success_count,
-            other_failure_count,
-            unknown_count,
-            updated_at
-        ) VALUES (?, ?, 86400, ?, ?, ?, ?, ?, ?, 0, 0, 0, ?)
-        "#,
-    )
-    .bind(key_id)
-    .bind(bucket_start)
-    .bind(total_requests)
-    .bind(success_count)
-    .bind(error_count)
-    .bind(quota_exhausted_count)
-    .bind(success_count)
-    .bind(error_count + quota_exhausted_count)
-    .bind(bucket_start + 60)
-    .execute(&proxy.key_store.pool)
-    .await
-    .expect("insert summary window bucket");
-}
-
-async fn insert_dashboard_summary_rollup_day_bucket(
-    proxy: &TavilyProxy,
-    bucket_start: i64,
-    total_requests: i64,
-    success_count: i64,
-    error_count: i64,
-    quota_exhausted_count: i64,
-) {
-    sqlx::query(
-        r#"
-        INSERT INTO dashboard_request_rollup_buckets (
-            bucket_start,
-            bucket_secs,
-            total_requests,
-            success_count,
-            error_count,
-            quota_exhausted_count,
-            valuable_success_count,
-            valuable_failure_count,
-            valuable_failure_429_count,
-            other_success_count,
-            other_failure_count,
-            unknown_count,
-            mcp_non_billable,
-            mcp_billable,
-            api_non_billable,
-            api_billable,
-            local_estimated_credits,
-            updated_at
-        ) VALUES (?, 86400, ?, ?, ?, ?, ?, ?, 0, 0, 0, 0, 0, 0, 0, ?, 0, ?)
-        "#,
-    )
-    .bind(bucket_start)
-    .bind(total_requests)
-    .bind(success_count)
-    .bind(error_count)
-    .bind(quota_exhausted_count)
-    .bind(success_count)
-    .bind(error_count + quota_exhausted_count)
-    .bind(total_requests)
-    .bind(bucket_start + 60)
-    .execute(&proxy.key_store.pool)
-    .await
-    .expect("insert dashboard summary rollup day bucket");
-}
-
-async fn insert_summary_window_logs(
-    proxy: &TavilyProxy,
-    key_id: &str,
-    created_at: i64,
-    outcome: &str,
-    count: usize,
-) {
-    insert_summary_window_logs_with_visibility(
-        proxy,
-        key_id,
-        created_at,
-        outcome,
-        count,
-        REQUEST_LOG_VISIBILITY_VISIBLE,
-    )
-    .await;
-}
-
-async fn insert_summary_window_logs_with_visibility(
-    proxy: &TavilyProxy,
-    key_id: &str,
-    created_at: i64,
-    outcome: &str,
-    count: usize,
-    visibility: &str,
-) {
-    for offset in 0..count {
-        sqlx::query(
-            r#"
-            INSERT INTO request_logs (
-                api_key_id,
-                auth_token_id,
-                method,
-                path,
-                query,
-                status_code,
-                tavily_status_code,
-                error_message,
-                result_status,
-                request_body,
-                response_body,
-                forwarded_headers,
-                dropped_headers,
-                visibility,
-                created_at
-            ) VALUES (?, NULL, 'GET', '/api/tavily/search', NULL, 200, 200, NULL, ?, NULL, NULL, '[]', '[]', ?, ?)
-            "#,
-        )
-        .bind(key_id)
-        .bind(outcome)
-        .bind(visibility)
-        .bind(created_at + offset as i64)
-        .execute(&proxy.key_store.pool)
-        .await
-        .expect("insert summary window log");
-    }
-
-    proxy
-        .key_store
-        .rebuild_dashboard_request_rollup_buckets_window(
-            Some(created_at),
-            Some(created_at + count as i64),
-        )
-        .await
-        .expect("rebuild dashboard rollup after summary log seed");
-}
-
-#[derive(Clone)]
-struct DashboardHourlyLogSeed<'a> {
-    created_at: i64,
-    path: &'a str,
-    request_kind_key: &'a str,
-    request_kind_label: &'a str,
-    result_status: &'a str,
-    failure_kind: Option<&'a str>,
-    request_body: Option<&'a [u8]>,
-    visibility: &'a str,
-}
-
-async fn insert_dashboard_hourly_log(
-    proxy: &TavilyProxy,
-    key_id: &str,
-    seed: DashboardHourlyLogSeed<'_>,
-) {
-    let status_code = match seed.result_status {
-        OUTCOME_SUCCESS => 200,
-        OUTCOME_QUOTA_EXHAUSTED => 429,
-        _ => 500,
-    };
-    let tavily_status_code = if seed.failure_kind == Some(FAILURE_KIND_UPSTREAM_RATE_LIMITED_429)
-        || seed.result_status == OUTCOME_QUOTA_EXHAUSTED
-    {
-        Some(429)
-    } else if seed.result_status == OUTCOME_SUCCESS {
-        Some(200)
-    } else {
-        Some(500)
-    };
-
-    sqlx::query(
-        r#"
-        INSERT INTO request_logs (
-            api_key_id,
-            auth_token_id,
-            method,
-            path,
-            query,
-            status_code,
-            tavily_status_code,
-            error_message,
-            result_status,
-            request_kind_key,
-            request_kind_label,
-            request_body,
-            response_body,
-            forwarded_headers,
-            dropped_headers,
-            visibility,
-            failure_kind,
-            created_at
-        ) VALUES (?, NULL, 'POST', ?, NULL, ?, ?, NULL, ?, ?, ?, ?, NULL, '[]', '[]', ?, ?, ?)
-        "#,
-    )
-    .bind(key_id)
-    .bind(seed.path)
-    .bind(status_code)
-    .bind(tavily_status_code)
-    .bind(seed.result_status)
-    .bind(seed.request_kind_key)
-    .bind(seed.request_kind_label)
-    .bind(seed.request_body)
-    .bind(seed.visibility)
-    .bind(seed.failure_kind)
-    .bind(seed.created_at)
-    .execute(&proxy.key_store.pool)
-    .await
-    .expect("insert dashboard hourly log");
-
-    proxy
-        .key_store
-        .rebuild_dashboard_request_rollup_buckets_window(
-            Some(seed.created_at),
-            Some(seed.created_at + 1),
-        )
-        .await
-        .expect("rebuild dashboard rollup after hourly log seed");
-}
-
-async fn insert_summary_window_charged_logs(
-    proxy: &TavilyProxy,
-    key_id: &str,
-    created_at: i64,
-    credits: i64,
-    count: usize,
-) {
-    for offset in 0..count {
-        sqlx::query(
-            r#"
-            INSERT INTO request_logs (
-                api_key_id,
-                auth_token_id,
-                method,
-                path,
-                query,
-                status_code,
-                tavily_status_code,
-                error_message,
-                result_status,
-                business_credits,
-                request_body,
-                response_body,
-                forwarded_headers,
-                dropped_headers,
-                visibility,
-                created_at
-            ) VALUES (?, NULL, 'GET', '/api/tavily/search', NULL, 200, 200, NULL, ?, ?, NULL, NULL, '[]', '[]', ?, ?)
-            "#,
-        )
-        .bind(key_id)
-        .bind(OUTCOME_SUCCESS)
-        .bind(credits)
-        .bind(REQUEST_LOG_VISIBILITY_VISIBLE)
-        .bind(created_at + offset as i64)
-        .execute(&proxy.key_store.pool)
-        .await
-        .expect("insert summary window charged log");
-    }
-
-    proxy
-        .key_store
-        .rebuild_dashboard_request_rollup_buckets_window(
-            Some(created_at),
-            Some(created_at + count as i64),
-        )
-        .await
-        .expect("rebuild dashboard rollup after charged log seed");
-}
-
-async fn insert_summary_window_maintenance_record(
-    proxy: &TavilyProxy,
-    key_id: &str,
-    created_at: i64,
-    source: &str,
-    operation_code: &str,
-    reason_code: Option<&str>,
-) {
-    let reason_summary = reason_code.map(|code| format!("{code} summary"));
-    sqlx::query(
-        r#"
-        INSERT INTO api_key_maintenance_records (
-            id,
-            key_id,
-            source,
-            operation_code,
-            operation_summary,
-            reason_code,
-            reason_summary,
-            reason_detail,
-            request_log_id,
-            auth_token_log_id,
-            auth_token_id,
-            actor_user_id,
-            actor_display_name,
-            status_before,
-            status_after,
-            quarantine_before,
-            quarantine_after,
-            created_at
-        ) VALUES (?, ?, ?, ?, ?, ?, ?, NULL, NULL, NULL, NULL, NULL, NULL, 'active', 'exhausted', 0, 0, ?)
-        "#,
-    )
-    .bind(format!("summary-window-maint-{}", nanoid!(8)))
-    .bind(key_id)
-    .bind(source)
-    .bind(operation_code)
-    .bind(format!("{operation_code} summary"))
-    .bind(reason_code)
-    .bind(reason_summary)
-    .bind(created_at)
-    .execute(&proxy.key_store.pool)
-    .await
-    .expect("insert summary window maintenance record");
-}
-
-async fn insert_auth_token_metric_log(
-    proxy: &TavilyProxy,
-    token_id: &str,
-    created_at: i64,
-    result_status: &str,
-) {
-    sqlx::query(
-        r#"
-        INSERT INTO auth_token_logs (
-            token_id,
-            method,
-            path,
-            query,
-            http_status,
-            mcp_status,
-            result_status,
-            error_message,
-            created_at,
-            counts_business_quota
-        ) VALUES (?, 'POST', '/api/tavily/search', NULL, 200, 200, ?, NULL, ?, 1)
-        "#,
-    )
-    .bind(token_id)
-    .bind(result_status)
-    .bind(created_at)
-    .execute(&proxy.key_store.pool)
-    .await
-    .expect("insert auth token metric log");
-}
-
 #[tokio::test]
 async fn summary_windows_split_today_yesterday_and_month() {
     let db_path = temp_db_path("summary-windows-split");
@@ -2133,14 +1773,7 @@ async fn summary_windows_split_today_yesterday_and_month() {
     )
     .await;
     insert_summary_window_logs(&proxy, &key_id, yesterday_same_time, OUTCOME_SUCCESS, 1).await;
-    insert_summary_window_logs(
-        &proxy,
-        &key_id,
-        yesterday_same_time + 1,
-        OUTCOME_SUCCESS,
-        3,
-    )
-    .await;
+    insert_summary_window_logs(&proxy, &key_id, yesterday_same_time + 1, OUTCOME_SUCCESS, 3).await;
 
     let mut expected_month = SummaryWindowMetrics {
         total_requests: 12,
